+++ src/libYARP_OS/src/MultiNameSpace.cpp
@@ -115,7 +115,7 @@ public:
         }
         Bottle ns = conf.getNamespaces();
         // loop through namespaces
-        for (int i=0; i<ns.size(); i++) {
+        for (size_t i=0; i<ns.size(); i++) {
             std::string n = ns.get(i).asString();
             NameConfig conf2;
             // read configuration of individual namespace
