+++ src/yarpmanager/src-manager/main.cpp
@@ -132,10 +132,10 @@ int main(int argc, char *argv[])
 //        yarp::os::Bottle appPaths2=rf.findPaths("scripts");
 //        std::cout << "app path : " << appPaths.toString()<< std::endl;
         QString appPathsStr="";
-        for (int ind=0; ind < appPaths.size(); ++ind){
+        for (size_t ind=0; ind < appPaths.size(); ++ind){
             appPathsStr += (appPaths.get(ind).asString() + ";").c_str();
         }
-        for (int ind=0; ind < appPaths2.size(); ++ind){
+        for (size_t ind=0; ind < appPaths2.size(); ++ind){
             appPathsStr += (appPaths2.get(ind).asString() + ";").c_str();
         }
         if (add_curr_dir)
@@ -150,7 +150,7 @@ int main(int argc, char *argv[])
        appPaths=rf.findPaths("modules");
        //std::cout << "mod path : " << appPaths.toString()<< std::endl;
        QString modPathsStr="";
-       for (int ind=0; ind < appPaths.size(); ++ind){
+       for (size_t ind=0; ind < appPaths.size(); ++ind){
            modPathsStr += (appPaths.get(ind).asString() + ";").c_str();
        }
        if (add_curr_dir)
@@ -164,7 +164,7 @@ int main(int argc, char *argv[])
        appPaths=rf.findPaths("resources");
        //std::cout << "res path : " << appPaths.toString()<< std::endl;
        QString resPathsStr="";
-       for (int ind=0; ind < appPaths.size(); ++ind){
+       for (size_t ind=0; ind < appPaths.size(); ++ind){
            resPathsStr += (appPaths.get(ind).asString() + ";").c_str();
        }
        if (add_curr_dir)
@@ -178,7 +178,7 @@ int main(int argc, char *argv[])
        appPaths=rf.findPaths("templates/applications");
       // std::cout << "templ path : " << appPaths.toString()<< std::endl;
        QString templPathsStr="";
-       for (int ind=0; ind < appPaths.size(); ++ind){
+       for (size_t ind=0; ind < appPaths.size(); ++ind){
             templPathsStr += (appPaths.get(ind).asString() + ";").c_str();
        }
        if (add_curr_dir)
